moving avatar / cleaning files
