+++ src/components/Header/NavBar.js
@@ -3,7 +3,7 @@ import Image from "next/image";
 import ReusableNavLink from "../common/ReusableNavLink";
 import { useCollection } from "@/hook/useCollection";
 
-export default function NavBar({ query, handleInputChange, data, fetchData }) {
+export default function NavBar({ fetchData }) {
   const navList = [
     { text: "About", href: "/about" },
     { text: "Contact", href: "/contact" },
@@ -52,13 +52,6 @@ export default function NavBar({ query, handleInputChange, data, fetchData }) {
             required
             onChange={fetchData}
           />
-          {/* <button
-            type="submit"
-            className="absolute right-1 top-1/2 transform -translate-y-1/2 bg-gradient-to-r from-gradient-lite-blue to-gradient-dark-blue text-white py-2 px-6 rounded-xl hover:bg-gray-500 active:bg-gray-600 focus:outline-none focus:ring-2 focus:ring-blue-300"
-            onClick={() => fetchData(query, data)}
-          >
-            SEARCH
-          </button> */}
         </div>
       </div>
     </div>
