--- src/components/Header/NavBar.module.css
@@ -1,9 +1,0 @@
-
-
-.nav {
-    list-style-type: none;
-    display: flex;
-    gap: 1rem;
-}
-
-
