+++ src/components/common/CustomizableAvatar.js
