+++ src/pages/managerDashboard.js
@@ -1,15 +1,13 @@
-import React from 'react';
-import Link from 'next/link';
-import NavBar from '@/components/Header/NavBar';
-import CustomizableAvatar from '@/components/common/customavatar/CustomizableAvatar';
-import CustomizableButton from '@/components/common/CustomizableButton';
-import FooterBar from '@/components/common/FooterBar';
-import styles from '../styles/singleListing.module.css';
+import React from "react";
+import Link from "next/link";
+import NavBar from "@/components/Header/NavBar";
+import CustomizableAvatar from "@/components/common/CustomizableAvatar";
+import CustomizableButton from "@/components/common/CustomizableButton";
+import FooterBar from "@/components/common/FooterBar";
+import styles from "../styles/singleListing.module.css";
 
 const managerDashboard = () => {
-  return (
-    <div>managerDashboard</div>
-  )
-}
+  return <div>managerDashboard</div>;
+};
 
-export default managerDashboard
+export default managerDashboard;
